third_party/WebKit/LayoutTests/web-animations-api/w3c/get-css-players.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script><style>
@keyframes test {
    from { opacity: 0; }
    to { opacity: 1; }
}
.cssAnimation {
    animation: test 2s;
}
</style>
<div id='container'>
    <div id='element'></div>
</div>

<script>
async_test(function(t) {
    assert_equals(document.timeline.getAnimationPlayers().length, 0);
    assert_equals(container.getAnimationPlayers().length, 0);
    assert_equals(element.getAnimationPlayers().length, 0);

    element.className = 'cssAnimation';
    onload = function () {
        t.step(function() {
            var players = document.timeline.getAnimationPlayers();
            assert_equals(players.length, 1);
            assert_equals(container.getAnimationPlayers().length, 0);
            assert_equals(element.getAnimationPlayers().length, 1);

            players[0].finish();
            assert_equals(document.timeline.getAnimationPlayers().length, 0);
            assert_equals(container.getAnimationPlayers().length, 0);
            assert_equals(element.getAnimationPlayers().length, 0);
            t.done();
        });
    }
}, 'getAnimationPlayers() with cssanimations');

</script>
